refactor: split poisson utils modules
